+++ CHANGELOG.md
@@ -119,6 +119,7 @@
     * Renamed arguments `alpha_perimeter` to `alpha_line`, `color_perimeter` to `color_line` to align with `LineStrings`.
     * Renamed arguments `alpha_fill` to `alpha_face` and `color_fill` to `color_face`.
 * Refactored `Polygon.project()` to use `LineString.project()`.
+* Refactored `Polygon.shift()` to use `LineString.shift()`.
 
 
 # 0.2.8